Bug fixing for multiplayer purpose.. 2.0 : changed housePlots and roadPlots from Arraylist to Arrays to prevent skipping index positions when removing available plots, now pre-allocates the specific size array and fills it with the positions..

+++ src/Network/ServerData.java
@@ -20,7 +20,7 @@ public class ServerData {
     public String[] textToRender = new String[]{"","",""};
     public String[] oldText = new String[10];
     public ArrayList<Integer> cards = new ArrayList<Integer>();
-    public float[] serializedHouse = new float[3];
+    public int[] serializedHouse = new int[3];
 
     Integer[] yieldNumbers = {2, 3, 3, 4, 4, 5, 5, 6, 6, 8, 8, 9, 9, 10, 10, 11, 11, 12};
     ArrayList<Integer> listOfYieldNumbers = new ArrayList<Integer>(Arrays.asList(yieldNumbers));
